Added: new features added to the app

+++ src/main/java/com/example/bookstore/util/TokenUtil.java
@@ -14,7 +14,7 @@ public class TokenUtil {
     private static final String TOKEN_SECRET = "BookStore";
 
 
-    public String createToken(long id)   {
+    public String createToken(Long id)   {
         try {
             //to set algorithm
             Algorithm algorithm = Algorithm.HMAC256(TOKEN_SECRET);
@@ -33,9 +33,9 @@ public class TokenUtil {
         }
         return null;
     }
-    public long decodeToken(String token)
+    public Long decodeToken(String token)
     {
-        long userid;
+        Long userid;
         //for verification algorithm
         Verification verification = null;
         try {
@@ -49,7 +49,7 @@ public class TokenUtil {
         DecodedJWT decodedjwt=jwtverifier.verify(token);
 
         Claim claim=decodedjwt.getClaim("user_id");
-        userid=claim.asInt();
+        userid=claim.asLong();
         return userid;
 
     }
